extracted QRCode detection to a new class
